refactor(tasks): extract display functions to prompts/display.ts
